pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "email-mcp-server"
version = "0.1.0"
description = "MCP Email Server for sending emails via SMTP"
authors = [{name = "dew", email = "anzil.thh@gmail.com"}]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "mcp>=1.0.0",
    "aiosmtplib>=2.0.2",
    "python-dotenv>=1.0.1",
]

[project.scripts]
email-mcp = "smtp_uvx_mcp.main:main"

[project.urls]
Homepage = "https://github.com/dewkiks/smtp-uvx-mcp"
Repository = "https://github.com/dewkiks/smtp-uvx-mcp"

[tool.hatch.build.targets.wheel]
packages = [
  { include = "smtp_uvx_mcp", from = "src" }
]